% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/async.R
\name{async}
\alias{async}
\title{Create an asynchroneous task}
\usage{
async(expr, envir = parent.frame())
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment from which global environment
             are search from.}
}
\value{
An AsyncTask object
}
\description{
Create an asynchroneous task
}

