% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/asyncEvalQ.R
\name{asyncEvalQ}
\alias{asyncEvalQ}
\title{Evaluate multiple R expressions asynchronously}
\usage{
asyncEvalQ(exprs, ..., envir = parent.frame())
}
\arguments{
\item{exprs}{A list of R expressions.}

\item{...}{Additional arguments passed to \code{asyncBatchEvalQ)}.}

\item{envir}{The environment from where to search for globals.}
}
\value{
A \code{\link{listenv}} of length \code{length(exprs)}.
}
\description{
Evaluate multiple R expressions asynchronously
}
\keyword{internal}

