% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyncListEnv.R
\name{AsyncListEnv}
\alias{AsyncListEnv}
\title{Create an async list environment}
\usage{
AsyncListEnv(...)
}
\arguments{
\item{\ldots}{Passed to \code{listenv()}.}
}
\value{
An environment of class `AsyncListEnv` extending `listenv`.
}
\description{
Create an async list environment
}

