% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{AsyncTask}
\alias{AsyncTask}
\alias{async}
\title{Create an asynchroneous task}
\usage{
AsyncTask(expr = NULL, envir = parent.frame(),
  finalize = getOption("async::finalize", TRUE))
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment in which global environment
should be located.}

\item{finalize}{If TRUE, any underlying registries are
deleted when this object is garbage collected, otherwise not.}
}
\value{
An AsyncTask object
}
\description{
Create an asynchroneous task
}
\keyword{internal}

