% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/availableCores.R
\name{availableCores}
\alias{availableCores}
\title{Get number of available cores on current machine}
\usage{
availableCores(sources = c("PBS", "system"))
}
\arguments{
\item{sources}{A character vector specifying which sources to use
   to infer the number of available cores (and in what order).}
}
\value{
Return a positive integer equal or greater to one.
}
\description{
Get number of available cores on current machine
}
\details{
The following sources for inferring the number of cores are supported:
\itemize{
 \item \code{"PBS"} -
   Query Torque/PBS environment variable \code{PBS_NUM_PPN}.
 \item \code{"system"} -
   Query \code{\link[parallel]{detectCores}()}.
}
}
\seealso{
\code{\link[parallel]{detectCores}()}
}

