% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AsyncResult.R
\name{await.AsyncResult}
\alias{await.AsyncResult}
\title{Retrieves the value of of the asynchronously evaluated expression}
\usage{
\method{await}{AsyncResult}(task, cleanup = FALSE, maxTries = 10L,
  interval = getOption("async::pollinterval", 1), ...)
}
\arguments{
\item{task}{The asynchronously task}

\item{cleanup}{If TRUE, the registry is completely removed upon
success, otherwise not.}

\item{maxTries}{The number of polls before giving up.}

\item{interval}{The number of seconds to wait between polls.}

\item{...}{Not used.}
}
\value{
The value of the evaluated expression.
If an error occurs, an informative Exception is thrown.
}
\description{
Retrieves the value of of the asynchronously evaluated expression
}

