% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsFuture-class.R
\name{await.BatchJobsFuture}
\alias{await.BatchJobsFuture}
\title{Retrieves the value of of the asynchronously evaluated expression}
\usage{
\method{await}{BatchJobsFuture}(future, cleanup = TRUE,
  maxTries = getOption("async::maxTries", Sys.getenv("R_ASYNC_MAXTRIES",
  1000)), delta = getOption("async::interval", 1), alpha = 1.01, ...)
}
\arguments{
\item{future}{The future.}

\item{cleanup}{If TRUE, the registry is completely removed upon
success, otherwise not.}

\item{maxTries}{The number of tries before giving up.}

\item{delta}{The number of seconds to wait the first time.}

\item{alpha}{A factor to scale up the waiting time in each iteration
such that the waiting time in the k:th iteration is \code{alpha^k*delta}.}

\item{\ldots}{Not used.}
}
\value{
The value of the evaluated expression.
If an error occurs, an informative Exception is thrown.
}
\description{
Retrieves the value of of the asynchronously evaluated expression
}
\keyword{internal}

