% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{await}
\alias{await}
\title{Retrieves the value of of the asynchronously evaluated expression}
\usage{
await(...)
}
\arguments{
\item{...}{Arguments passed to S3 method.}
}
\value{
The value of the evaluated expression.
If an error occurs, an informative Exception is thrown.
}
\description{
Retrieves the value of of the asynchronously evaluated expression
}
\keyword{internal}

