% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batchEvalQ.R
\name{batchEvalQ}
\alias{batchEvalQ}
\title{Evaluates expressions as jobs on a registry}
\usage{
batchEvalQ(reg, exprs, local = FALSE, ...)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Empty Registry that will store jobs for the mapping.}

\item{exprs}{[\code{\link{expression}}]\cr
R expressions to evaluate.}

\item{local}{[\code{\link{logical}}[1]]\cr
If TRUE, the expression is evaluated wrapped in a
\code{\link[base]{local}} call, otherwise directly
in the global environment (of the job).}

\item{...}{[any]\cr
  Additional arguments passed to \code{\link{batchMap}}.}
}
\value{
Vector of type \code{integer} with job ids.
}
\description{
Evaluates expressions as jobs on a registry
}
\seealso{
This function is analogous to
\code{parallel::\link[parallel]{clusterEvalQ}}.
}

