% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchjobs.R
\name{batchjobs}
\alias{batchjobs}
\title{A BatchJobs future represents a future that is resolved by one of the many BatchJobs backends}
\usage{
batchjobs(expr, envir = parent.frame(), substitute = TRUE, backend = NULL,
  finalize = getOption("async::finalize", TRUE), ...)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envir}{The environment from which global environment
are search from.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{...}{Additional arguments pass to \code{\link{BatchJobsAsyncTask}()}.}
}
\value{
Returns a BatchJobsAsyncTask object that also is
a \link[future]{Future}.
}
\description{
A BatchJobs future represents a future that is resolved by one of the many BatchJobs backends
}

