% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchjobs_conf.R
\name{batchjobs_conf}
\alias{batchjobs_conf}
\title{BatchJobs conf futures}
\usage{
batchjobs_conf(expr, envir = parent.frame(), substitute = TRUE, pathnames,
  ...)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envir}{The environment from which global environment
are search from.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{pathnames}{Pathnames to one or more BatchJobs configuration files
to be loaded in order.}

\item{\ldots}{Additional arguments passed to \code{\link{BatchJobsFuture}()}.}
}
\value{
An object of class \code{BatchJobsFuture}.
}
\description{
A conf BatchJobs future uses one or more BatchJobs configuration
files to define the BatchJobs configuration environment,
e.g. \file{.BatchJobs.R}.
}

