% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchjobs_custom.R
\name{batchjobs_custom}
\alias{batchjobs_custom}
\title{BatchJobs custom futures}
\usage{
batchjobs_custom(expr, envir = parent.frame(), substitute = TRUE, pathname,
  ...)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envir}{The environment from which global environment
are search from.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{pathname}{The pathname to a BatchJobs configuration file.}

\item{\ldots}{Additional arguments passed to \code{\link{BatchJobsFuture}()}.}
}
\value{
An object of class \code{BatchJobsFuture}.
}
\description{
A custom BatchJobs future uses a BatchJobs configuration file
that defines the BatchJobs configuration environment, e.g. \file{.BatchJobs.R}.
}

