% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchjobs_torque.R
\name{batchjobs_lsf}
\alias{batchjobs_lsf}
\alias{batchjobs_openlava}
\alias{batchjobs_sge}
\alias{batchjobs_slurm}
\alias{batchjobs_torque}
\title{BatchJobs LSF, OpenLava, SGE, Slurm and Torque futures}
\usage{
batchjobs_lsf(expr, envir = parent.frame(), substitute = TRUE,
  pathname = NULL, ...)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envir}{The environment from which global environment
are search from.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{pathname}{A BatchJobs template file (brew formatted).}

\item{\ldots}{Additional arguments passed to \code{\link{BatchJobsFuture}()}.}
}
\value{
An object of class \code{BatchJobsFuture}.
}
\description{
LSF, OpenLava, SGE, Slurm and Torque BatchJobs futures are asynchronous
and multiprocess, will not block, and will be evaluated on a compute
cluster via a job scheduler.
}

