% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchjobs_multicore.R
\name{batchjobs_multicore}
\alias{batchjobs_multicore}
\title{BatchJobs multicore futures}
\usage{
batchjobs_multicore(expr, envir = parent.frame(), substitute = TRUE,
  workers = availableCores(constraints = "multicore"), ...)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envir}{The environment from which global environment
are search from.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{workers}{The number of multicore processes to be available
for concurrent BatchJobs multicore futures.}

\item{\ldots}{Additional arguments passed to \code{\link{BatchJobsFuture}()}.}
}
\value{
An object of class \code{BatchJobsFuture}.
}
\description{
A multicore BatchJobs future is asynchronous and multiprocess,
will be evaluated immediately and will not block.
}
\seealso{
An alternative to BatchJobs multicore futures is to use
\link[future:multicore]{multicore} futures of the \pkg{future}.

Internally \code{\link[BatchJobs]{makeClusterFunctionsMulticore}()}
is used to create the BatchJobs cluster-function backend.
}

