% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BatchJobsAsyncTask.R
\name{BatchJobsAsyncTask}
\alias{BatchJobsAsyncTask}
\title{Create an asynchroneous task}
\usage{
BatchJobsAsyncTask(expr = NULL, envir = parent.frame(), substitute = TRUE,
  backend = NULL, finalize = getOption("async::finalize", TRUE),
  launch = TRUE, ...)
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment in which global environment
should be located.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{finalize}{If TRUE, any underlying registries are
deleted when this object is garbage collected, otherwise not.}
}
\value{
An AsyncTask object
}
\description{
Create an asynchroneous task
}
\keyword{internal}

