% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsFuture-class.R
\name{BatchJobsFuture}
\alias{BatchJobsFuture}
\title{A BatchJobs future is a future whose value will be resolved via BatchJobs}
\usage{
BatchJobsFuture(expr = NULL, envir = parent.frame(), substitute = TRUE,
  conf = NULL, cluster.functions = NULL, args = NULL, workers = NULL,
  finalize = getOption("future.finalize", TRUE), ...)
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment in which global environment
should be located.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{conf}{A BatchJobs configuration environment.}

\item{cluster.functions}{A BatchJobs \link[BatchJobs]{ClusterFunctions} object.}

\item{args}{Named list of additional arguments passed to the BatchJobs brew template, which will be available to the template as variables with the same name.  BatchJobs supports only \code{resources}, so any attempts to pass other variables will result in an immediate error.}

\item{workers}{(optional) Additional specification for
the BatchJobs backend.}

\item{finalize}{If TRUE, any underlying registries are
deleted when this object is garbage collected, otherwise not.}

\item{\ldots}{Additional arguments passed to \code{\link[future]{Future}()}.}
}
\value{
A BatchJobsFuture object
}
\description{
A BatchJobs future is a future whose value will be resolved via BatchJobs
}
\keyword{internal}

