% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsFutureError.R
\docType{data}
\name{BatchJobsFutureError}
\alias{BatchJobsFutureError}
\alias{getOutput}
\title{Exception class for BatchJobsFuture-related errors}
\format{An object of class \code{Class} (inherits from \code{Object}) of length 1.}
\usage{
BatchJobsFutureError
}
\arguments{
\item{future}{The BatchJobs future for which the error occurred.}

\item{\ldots}{Passed to \code{\link[R.oo]{Exception}}}
}
\description{
Exception class for BatchJobsFuture-related errors
}
\keyword{internal}

