% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delayedAsyncAssign.R
\name{delayedAsyncAssign}
\alias{\%<=\%}
\alias{\%=>\%}
\alias{delayedAsyncAssign}
\title{Delayed asynchroneous evaluation}
\usage{
delayedAsyncAssign(name, value, envir = parent.frame(), assign.env = envir,
  substitute = TRUE)
}
\arguments{
\item{name}{the name of the variable to assign.}

\item{value}{the R expression to be asynchroneous evaluated and
whose value will be assigned to the variable.}

\item{envir}{The environment from which global variables used by
the expression should be search for.}

\item{assign.env}{The environment to which the variable should
be assigned.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}
}
\value{
A delayed assignment which, when evaluated, will retrieve
the value of the asynchronous evaluation.
}
\description{
Method and infix operators for delayed assignments while evaluating
the statement in the background/in parallel.
}
\examples{
backend(c("multicore", "local"))

## A global variable
a <- 1

## Three asynchronous evaluations
A \%<=\% { x <- 0.1 }
B \%<=\% { y <- 0.2 }
C \%<=\% { z <- a+0.3 }

## Sleep until 'C' is available
message("Value 'C': ", C)

## Sleep until 'A' is available
message("Value 'A': ", A)

## Sleep until 'C' is available
message("Value 'B': ", B)
}
\seealso{
\code{\link{asyncEvalQ}()}
}

