% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{delete.AsyncTask}
\alias{delete}
\alias{delete.AsyncTask}
\title{Removes an asynchroneous task}
\usage{
\method{delete}{AsyncTask}(task, ...)
}
\arguments{
\item{task}{The asynchronously task}

\item{...}{Not used.}

\item{onFailure}{Action if failing to delete task.}

\item{onMissing}{Action if task does not exist.}

\item{maxTries}{The number of tries before giving up.}

\item{delta}{The delay interval (in seconds) between retries.}

\item{alpha}{A multiplicative penalty increasing the delay
for each failed try.}
}
\value{
(invisibly) TRUE if deleted and FALSE otherwise.
}
\description{
Removes an asynchroneous task
}
\keyword{internal}

