% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BatchJobsAsyncTask.R
\name{delete.BatchJobsAsyncTask}
\alias{delete.BatchJobsAsyncTask}
\title{Removes an asynchroneous task}
\usage{
\method{delete}{BatchJobsAsyncTask}(obj, onFailure = c("error", "warning",
  "ignore"), onMissing = c("ignore", "warning", "error"), maxTries = 10L,
  delta = getOption("async::interval", 1), alpha = 1.01, ...)
}
\arguments{
\item{obj}{The asynchronously task}

\item{onFailure}{Action if failing to delete task.}

\item{onMissing}{Action if task does not exist.}

\item{maxTries}{The number of tries before giving up.}

\item{delta}{The delay interval (in seconds) between retries.}

\item{alpha}{A multiplicative penalty increasing the delay
for each failed try.}

\item{...}{Not used.}
}
\value{
(invisibly) TRUE if deleted and FALSE otherwise.
}
\description{
Removes an asynchroneous task
}
\keyword{internal}

