% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{delete}
\alias{delete}
\alias{delete.AsyncTask}
\title{Removes an asynchroneous task}
\usage{
delete(...)
}
\arguments{
\item{task}{The asynchronously task}

\item{onRunning}{Action if task is running or appears to run.}

\item{onFailure}{Action if failing to delete task.}

\item{onMissing}{Action if task does not exist.}

\item{maxTries}{The number of tries before giving up.}

\item{delta}{The delay interval (in seconds) between retries.}

\item{alpha}{A multiplicative penalty increasing the delay
for each failed try.}

\item{\ldots}{Not used.}
}
\value{
(invisibly) TRUE if deleted and FALSE otherwise.
}
\description{
Removes an asynchroneous task
}
\keyword{internal}

