% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.BatchJobs-package.R
\docType{package}
\name{future.BatchJobs}
\alias{future.BatchJobs}
\alias{future.BatchJobs-package}
\title{future.BatchJobs: A Future for BatchJobs}
\description{
The \pkg{future.BatchJobs} package implements the Future API
on top of \pkg{BatchJobs} such that futures can be resolved
on for instance high-performance compute (HPC) clusters via
job schedulers.
The Future API is defined by the \pkg{future} package.
}
\details{
To use BatchJobs futures, load \pkg{future.BatchJobs}, and
select the type of future you wish to use via
\code{\link[future:plan]{plan}()}.
}
\examples{
## Use local BatchJobs futures
plan(batchjobs_local)

## A global variable
a <- 1

v \%<-\% {
  b <- 3
  c <- 2
  a * b * c
}

print(v)
\donttest{
plan(batchjobs_local)
demo("mandelbrot", package="future", ask=FALSE)
}

}
