% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/globals.R
\name{getGlobals}
\alias{getGlobals}
\title{Get all global objects for one or more R expressions}
\usage{
getGlobals(expr, envir = parent.frame(), ..., unlist = TRUE)
}
\arguments{
\item{expr}{An R expression or a a list of R expressions.}

\item{envir}{The environment where to search for globals.}

\item{...}{Not used.}

\item{unlist}{If TRUE, a list of unique objects is returned.
       If FALSE, a list of \code{length(expr)} sublists.}
}
\value{
A named list of variables.
}
\description{
Get all global objects for one or more R expressions
}

