% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delayedAssign_OP.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Delayed synchroneous evaluation ("lazy evaluation")}
\usage{
x %<-% value
}
\description{
Infix operator for delayed ("lazy") synchroneous evaluation
such that the right-hand side (RHS) expression is evaluated
if and only if the left-hand side (LHS) object is requested.
}
\seealso{
Internally \link[base]{delayedAssign}() is used.
}

