% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assigner.R
\name{\%<-|\%}
\alias{\%<-|\%}
\title{Delayed asynchroneous assignment}
\usage{
\%|\%(x) <- value
}
\description{
Operator for delayed assignments while evaluating
the statement in the background/in parallel.
}

