% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect an asynchroneous variable}
\usage{
inspect(var = NULL, envir = parent.frame(), inherits = TRUE)
}
\arguments{
\item{var}{the variable.}

\item{envir}{the environment where to search from.}

\item{inherits}{Search parent frames or not.}
}
\value{
If exists, an AsyncTask object, otherwise NA.
}
\description{
Inspect an asynchroneous variable
}

