% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbrOfWorkers.R
\name{nbrOfWorkers.batchjobs}
\alias{nbrOfWorkers.batchjobs}
\alias{nbrOfWorkers.batchjobs_interactive}
\alias{nbrOfWorkers.batchjobs_multicore}
\title{Gets the number of BatchJobs workers}
\usage{
\method{nbrOfWorkers}{batchjobs}(evaluator)
}
\arguments{
\item{evaluator}{A future evaluator function.
If NULL (default), the current evaluator as returned
by \code{\link{plan}()} is used.}
}
\value{
A number in [1,Inf].
}
\description{
Tries to infer the total number of BatchJobs workers.  This is
done using various ad hoc procedures based on code inspection
of BatchJobs itself.
}
\keyword{internal}
