% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseEnvSubset.R
\name{parseEnvSubset}
\alias{parseEnvSubset}
\title{Helper function to infer target from expression and environment}
\usage{
parseEnvSubset(expr, envir = parent.frame(), substitute = TRUE)
}
\arguments{
\item{expr}{An expression.}

\item{envir}{An environment.}

\item{substitute}{If TRUE, then the expression is
       \code{substitute()}:ed, otherwise not.}
}
\value{
A named list.
}
\description{
Helper function to infer target from expression and environment
}
\keyword{internal}

