% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsFuture-class.R
\name{run.BatchJobsFuture}
\alias{run.BatchJobsFuture}
\title{Starts an asynchronously task}
\usage{
\method{run}{BatchJobsFuture}(task, ...)
}
\arguments{
\item{task}{The asynchronously task.}

\item{\ldots}{Not used.}
}
\description{
Starts an asynchronously task
}
\keyword{internal}

