% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{run}
\alias{run}
\title{Starts an asynchroneous task}
\usage{
run(task, ...)
}
\arguments{
\item{task}{The asynchronously task.}

\item{\ldots}{Not used.}
}
\value{
The the started task.
}
\description{
Starts an asynchroneous task
}
\keyword{internal}

