% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BatchJobs,AsyncTask.R
\name{status.BatchJobsAsyncTask}
\alias{status.BatchJobsAsyncTask}
\title{Status of an AsyncTask}
\usage{
\method{status}{BatchJobsAsyncTask}(obj, ...)
}
\arguments{
\item{obj}{The asynchronously task}

\item{...}{Not used.}
}
\value{
A character vector.
}
\description{
Status of an AsyncTask
}
\keyword{internal}

