% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{status}
\alias{completed}
\alias{error}
\alias{expired}
\alias{failed}
\alias{finished}
\alias{status}
\alias{value}
\title{Status of an AsyncTask}
\usage{
status(...)
}
\arguments{
\item{\ldots}{Arguments passed to the S3 method}
}
\value{
A logical or aA character vector.
}
\description{
Status of an AsyncTask
}
\keyword{internal}

