% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{task}
\alias{task}
\title{Get an asynchroneous expression}
\usage{
task(var, envir = parent.frame(), inherits = TRUE)
}
\arguments{
\item{var}{the asynchroneous variable.}

\item{envir}{the environment where to search from.}

\item{inherits}{Search parent frames or not.}
}
\value{
An AsyncTask object
}
\description{
Get an asynchroneous expression
}

