% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/async.R
\name{async}
\alias{async}
\title{Evaluate an expression asynchronously}
\usage{
async(expr, envir = parent.frame())
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment in which to expression should be evaluated.}
}
\value{
An AsyncResult
}
\description{
Evaluate an expression asynchronously
}

