% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/async-await.R
\name{async}
\alias{async}
\title{Explicit asynchroneous evaluation}
\usage{
async(expr, envir = parent.frame(), substitute = TRUE, ...)
}
\arguments{
\item{expr}{An R expression to be evaluated}

\item{envir}{The environment from which global environment
are search from.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{...}{Optional arguments.}
}
\value{
\code{async()} returns an AsyncTask object.
}
\description{
\code{async()} evaluates an expression asynchroneous.
}

