% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/async-await.R
\name{async}
\alias{async}
\alias{await}
\title{Explicit asynchroneous evaluation}
\usage{
async(expr, envir = parent.frame(), substitute = TRUE)

await(task, ...)
}
\arguments{
\item{expr}{An R expression to be evaluated}

\item{envir}{The environment from which global environment
are search from.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{task}{An \code{AsyncTask} object whose value to retrieve.}

\item{...}{Not used.}
}
\value{
\code{async()} returns an AsyncTask object.

\code{await()} returns the value of the expression.
}
\description{
\code{async()} evaluates an expression asynchroneous and
\code{await()} retrieves its value.
}

