% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BatchJobs,asyncBatchEvalQ.R
\name{asyncBatchEvalQ}
\alias{asyncBatchEvalQ}
\title{A batch map function for R expressions}
\usage{
asyncBatchEvalQ(reg, exprs, globals = TRUE, pkgs = NULL,
  envir = parent.frame(), ...)
}
\arguments{
\item{reg}{A Registry.}

\item{exprs}{A list of R expressions.}

\item{globals}{A named list of R objects to be loaded by each job.
If TRUE, globals are automatically searched for.}

\item{pkgs}{A character vector of package names to use.
This set will be extended with packages inferred from
the list of globals.}

\item{envir}{The environment where to search for globals.}

\item{...}{Additional arguments passed to \code{batchEvalQ)}.}
}
\value{
Vector of type \code{integer} with job ids.
}
\description{
A batch map function for R expressions
}
\keyword{internal}

