% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AsyncError.R
\docType{data}
\name{AsyncError}
\alias{AsyncError}
\alias{AsyncTaskError}
\title{Exception class for async-related errors}
\format{\preformatted{function (...)  
 - attr(*, ".env")=<environment: 0x000000000c7d22c0> 
 - attr(*, "class")= chr [1:2] "Class" "Object"
 - attr(*, "export")= logi TRUE
 - attr(*, "modifiers")= chr [1:2] "public" "class"
}}
\usage{
AsyncError(...)
}
\arguments{
\item{...}{Passed to \code{\link[R.oo]{Exception}}}
}
\description{
Exception class for async-related errors
}
\keyword{internal}

