% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{AsyncTask}
\alias{AsyncTask}
\title{Create an asynchroneous task}
\usage{
AsyncTask(expr = NULL, envir = parent.frame(), substitute = TRUE, ...)
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment in which global environment
should be located.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}
}
\value{
An AsyncTask object
}
\description{
Create an asynchroneous task
}
\keyword{internal}

