% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{AsyncTask}
\alias{AsyncTask}
\title{A asynchroneous task}
\usage{
AsyncTask(expr = NULL, reg = NA, id = 1L, ...)
}
\arguments{
\item{expr}{A R expression.}

\item{reg}{A BatchJob Registry}

\item{id}{A BatchJob Registry ID.}

\item{...}{Not used.}
}
\value{
An AsyncTask object
}
\description{
A asynchroneous task
}

