% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{await.AsyncTask}
\alias{await.AsyncTask}
\title{Retrieves the value of of the asynchronously evaluated expression}
\usage{
\method{await}{AsyncTask}(task, ...)
}
\arguments{
\item{task}{The asynchronously task}

\item{...}{Not used.}

\item{cleanup}{If TRUE, the registry is completely removed upon
success, otherwise not.}

\item{maxTries}{The number of tries before giving up.}

\item{delta}{The number of seconds to wait the first time.}

\item{alpha}{A factor to scale up the waiting time in each iteration
such that the waiting time in the k:th iteration is \code{alpha^k*delta}.}
}
\value{
The value of the evaluated expression.
If an error occurs, an informative Exception is thrown.
}
\description{
Retrieves the value of of the asynchronously evaluated expression
}
\keyword{internal}

