% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/async-await.R
\name{await}
\alias{await}
\title{Awaits an asynchroneous evaluation}
\usage{
await(task, ...)
}
\arguments{
\item{task}{An \code{AsyncTask} object whose value to retrieve.}

\item{...}{Not used.}
}
\value{
\code{await()} returns the value of the expression.
}
\description{
Awaits an asynchroneous evaluation
}
\details{
Note that \code{await()} should only be called once, because
after being called the actual asynchroneous task is removed
and no longer available for subsequent calls.  If called
again, an error is thrown.
}

