% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backend.R
\name{backend}
\alias{backend}
\title{Switch backend to be used for asynchronous processing}
\usage{
backend(what = c(".BatchJobs.R", "multicore-1", "multicore", "interactive",
  "local", "rscript"), ..., quietly = TRUE)
}
\arguments{
\item{what}{A character vector of preferred backend to be used.}

\item{...}{Named character arguments specifying custom aliases
for character sets of backends.}

\item{quietly}{If TRUE, messages are supressed.}
}
\value{
Returns the name of the backend used.
}
\description{
Switch backend to be used for asynchronous processing
}
\note{
The Windows operating system does not support the 'multicore'
backend, which then will be ignored.  If explicitly specified, then
an informative warning will be given.
}

