% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.backend.R
\name{backend}
\alias{backend}
\title{Switch backend to be used for asynchronous processing}
\usage{
backend(what = NULL, ..., quietly = TRUE)
}
\arguments{
\item{what}{A \code{character} \code{vector} of preferred backends to be used.
See Section Details below for supported backends.
If \code{NULL} (default), the currently default backend is returned.
If \code{"reset"}, the backend is reset to \code{"default"}.
If \code{"default"}, the default backend according to alias
   \code{"default"} is used (see Details).
If \code{"aliases"}, all registered aliases are returned.}

\item{quietly}{If TRUE, messages are suppressed.}

\item{\ldots}{Named character arguments specifying custom aliases
for character sets of backends.}
}
\value{
Returns the name of the backend used, or a list of named aliases.
}
\description{
\emph{This function is defunct.  Please use \code{plan()} with one
of the available \code{batchjobs_*} future strategies.}
}
\keyword{internal}
