% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batchEvalQ.R
\name{batchEvalQ}
\alias{batchEvalQ}
\title{A batch map function for R expressions}
\usage{
batchEvalQ(reg, exprs, globals = TRUE, envir = parent.frame(), ...)
}
\arguments{
\item{reg}{A Registry.}

\item{exprs}{A list of R expressions.}

\item{globals}{A named list of R objects to be loaded by each job.
If TRUE, globals are automatically searched for.}

\item{envir}{The environment where to search for globals.}

\item{...}{Additional arguments passed to \code{batchMap()}.}
}
\value{
Job IDs.
}
\description{
A batch map function for R expressions
}
\seealso{
This function is \pkg{BatchJobs}'s analogy to
\code{parallel::clusterEvalQ()} and \code{parallel::clusterExport()}
if globals are passed along as well.
}

