% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchjobs.R
\name{batchjobs}
\alias{batchjobs}
\title{A BatchJobs future represents a future that is resolved by one of the many BatchJobs backends}
\usage{
batchjobs(expr, envir = parent.frame(), substitute = TRUE, backend = NULL,
  resources = list(), finalize = getOption("async::finalize", TRUE), ...)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envir}{The environment from which global environment
are search from.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{backend}{The BatchJobs backend to use, cf. \code{\link{backend}()}.}

\item{resources}{A named list of resources needed by this future.}

\item{finalize}{If TRUE, any underlying registries are
deleted when this object is garbage collected, otherwise not.}

\item{\ldots}{Additional arguments pass to \code{\link{AsyncTask}()}.}
}
\value{
Returns a BatchJobsAsyncTask object that also is
a \link[future]{Future}.
}
\description{
A BatchJobs future represents a future that is resolved by one of the many BatchJobs backends
}

