% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.batchjobs.R
\name{batchjobs}
\alias{batchjobs}
\title{A BatchJobs future represents a future that is resolved by one of the many BatchJobs backends}
\usage{
batchjobs(expr, envir = parent.frame(), substitute = TRUE, backend = NULL,
  workers = NULL, ...)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envir}{The environment from which global environment
are search from.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{backend}{The BatchJobs backend to use.}

\item{workers}{(optional) Additional specification for the backend
workers.  If NULL, the default is used.}

\item{\ldots}{Additional arguments passed to \code{\link{BatchJobsFuture}()}.}
}
\value{
An object of class \code{BatchJobsFuture}.
}
\description{
\emph{This function is defunct.  Please use \code{plan()} with one
of the available \code{batchjobs_*} future strategies.}
}
\keyword{internal}
