% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchjobs_conf.R
\name{batchjobs_conf}
\alias{batchjobs_conf}
\title{BatchJobs conf futures}
\usage{
batchjobs_conf(expr, envir = parent.frame(), substitute = TRUE,
  pathnames = NULL, ...)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envir}{The environment from which global environment
are search from.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{pathnames}{Pathnames to one or more BatchJobs configuration
files to be loaded in order.  If NULL, then the \pkg{BatchJobs}
package will search for such configuration files.}

\item{\ldots}{Additional arguments passed to \code{\link{BatchJobsFuture}()}.}
}
\value{
An object of class \code{BatchJobsFuture}.
}
\description{
A conf BatchJobs future sources one or more BatchJobs configuration
files (R source scripts) to define the BatchJobs configuration
environment, e.g. \file{.BatchJobs.R}.
}
\details{
If \code{pathnames=NULL} (default), the \pkg{BatchJobs} is configured
to use (in order) each of following configuration files that exist:
\itemize{
 \item \code{system("etc", "BatchJobs_global_config.R", package="BatchJobs")}
 \item \code{~/.BatchJobs.R} (in user's home directory)
 \item \code{.BatchJobs.R} (in the current directory)
}
}

