% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsAsyncTask.R
\name{BatchJobsAsyncTask}
\alias{BatchJobsAsyncTask}
\title{Create an asynchroneous task}
\usage{
BatchJobsAsyncTask(expr = NULL, envir = parent.frame(), substitute = TRUE,
  backend = NULL, finalize = getOption("async::finalize", TRUE), ...)
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment in which global environment
should be located.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{backend}{The BatchJobs backend to use, cf. \code{\link{backend}()}.}

\item{finalize}{If TRUE, any underlying registries are
deleted when this object is garbage collected, otherwise not.}

\item{\ldots}{Additional arguments pass to \code{\link{AsyncTask}()}.}
}
\value{
A BatchJobsAsyncTask object
}
\description{
Create an asynchroneous task
}
\keyword{internal}

