% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsFuture-class.R
\name{BatchJobsFuture}
\alias{BatchJobsFuture}
\title{A BatchJobs future is a future whose value will be resolved via BatchJobs}
\usage{
BatchJobsFuture(expr = NULL, envir = parent.frame(), substitute = TRUE,
  backend = NULL, resources = list(),
  finalize = getOption("future.finalize", TRUE), ...)
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment in which global environment
should be located.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{backend}{The BatchJobs backend to use, cf. \code{\link{backend}()}.}

\item{resources}{A named list of resources needed by this future.}

\item{finalize}{If TRUE, any underlying registries are
deleted when this object is garbage collected, otherwise not.}

\item{\ldots}{Additional arguments pass to \code{\link[future]{Future}()}.}
}
\value{
A BatchJobsFuture object
}
\description{
A BatchJobs future is a future whose value will be resolved via BatchJobs
}
\keyword{internal}

