% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsFuture-class.R
\name{BatchJobsFuture}
\alias{BatchJobsFuture}
\title{A BatchJobs future is a future whose value will be resolved via BatchJobs}
\usage{
BatchJobsFuture(expr = NULL, envir = parent.frame(), substitute = TRUE,
  conf = NULL, cluster.functions = NULL, resources = list(),
  workers = NULL, finalize = getOption("future.finalize", TRUE), ...)
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment in which global environment
should be located.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{conf}{A BatchJobs configuration environment.}

\item{cluster.functions}{A BatchJobs \link[BatchJobs]{ClusterFunctions} object.}

\item{resources}{A named list passed to the BatchJobs template (available as variable `resources`).}

\item{workers}{(optional) Additional specification for
the BatchJobs backend.}

\item{finalize}{If TRUE, any underlying registries are
deleted when this object is garbage collected, otherwise not.}

\item{\ldots}{Additional arguments passed to \code{\link[future]{Future}()}.}
}
\value{
A BatchJobsFuture object
}
\description{
A BatchJobs future is a future whose value will be resolved via BatchJobs
}
\keyword{internal}

