% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsFutureError.R
\name{BatchJobsFutureError}
\alias{BatchJobsFutureError}
\title{FutureError class for errors related to BatchJobsFuture:s}
\usage{
BatchJobsFutureError(...)
}
\arguments{
\item{\ldots}{Arguments passed to \code{\link[future]{FutureError}}}
}
\description{
FutureError class for errors related to BatchJobsFuture:s
}
\keyword{internal}

