% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/delayedAsyncAssign.R
\name{delayedAsyncAssign}
\alias{\%<=\%}
\alias{\%=>\%}
\alias{delayedAsyncAssign}
\title{Delayed asynchroneous evaluation}
\usage{
delayedAsyncAssign(name, expr, assign.env = parent.frame(1))
}
\arguments{
\item{name}{the name of the variable to assign.}

\item{expr}{the R expression to be asynchroneous evaluated and
whose value will be assigned to the variable.}

\item{assign.env}{The environment to which the variable should
be assigned.}
}
\value{
A delayed assignment which, when evaluated, will retrieve
the value of the asynchronous evaluation.
}
\description{
Operator for delayed assignments while evaluating
the statement in the background/in parallel.
}
\seealso{
\code{\link{async}()}
}

