% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backend_OP.R, R/delayedAsyncAssign_OP.R
\name{\%backend\%}
\alias{\%backend\%}
\title{Evaluate asynchroneous expression on a specific backend}
\usage{
x %<-% { expr } %backend% backend

x %<-% { expr } %backend% backend
}
\description{
Evaluate asynchroneous expression on a specific backend

Evaluate asynchroneous expression on a specific backend
}

