% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect an asynchroneous variable}
\usage{
inspect(var = NULL, envir = parent.frame(), mustExist = FALSE)
}
\arguments{
\item{var}{the variable.}

\item{envir}{the environment where to search from.}

\item{mustExist}{If TRUE and the variable does not exists, then
an informative error is thrown, otherwise NA is returned.}
}
\value{
An AsyncTask object (or NA).
}
\description{
Inspect an asynchroneous variable
}

