% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsFuture-class.R
\name{print.BatchJobsFuture}
\alias{print.BatchJobsFuture}
\title{Prints a BatchJobs future}
\usage{
\method{print}{BatchJobsFuture}(x, ...)
}
\arguments{
\item{x}{An BatchJobsFuture object}

\item{\ldots}{Not used.}
}
\description{
Prints a BatchJobs future
}
\keyword{internal}

