% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsFuture-class.R
\name{run.BatchJobsFuture}
\alias{run.BatchJobsFuture}
\title{Runs a BatchJobs future}
\usage{
\method{run}{BatchJobsFuture}(future, ...)
}
\arguments{
\item{future}{The future.}

\item{\ldots}{Not used.}
}
\value{
(invisibly) the future.
}
\description{
Runs a BatchJobs future
}
\keyword{internal}

