% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{status.AsyncTask}
\alias{status.AsyncTask}
\title{Status of an AsyncTask}
\usage{
\method{status}{AsyncTask}(obj, ...)
}
\arguments{
\item{obj}{The asynchronously task}

\item{...}{Not used.}
}
\value{
A character vector.
}
\description{
Status of an AsyncTask
}
\keyword{internal}

