% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsAsyncTask.R
\name{status.BatchJobsAsyncTask}
\alias{status.BatchJobsAsyncTask}
\title{Status of an AsyncTask}
\usage{
\method{status}{BatchJobsAsyncTask}(task, ...)
}
\arguments{
\item{task}{The asynchronously task}

\item{\ldots}{Not used.}
}
\value{
A character vector.
}
\description{
Status of an AsyncTask
}
\keyword{internal}

