% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchJobsFuture-class.R
\name{status.BatchJobsFuture}
\alias{completed}
\alias{error}
\alias{expired}
\alias{failed}
\alias{finished}
\alias{status}
\alias{status.BatchJobsFuture}
\alias{value}
\title{Status of BatchJobs future}
\usage{
\method{status}{BatchJobsFuture}(future, ...)
}
\arguments{
\item{future}{The future.}

\item{\ldots}{Not used.}
}
\value{
A character vector or a logical scalar.
}
\description{
Status of BatchJobs future
}
\keyword{internal}

