% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AsyncTask.R
\name{status}
\alias{error}
\alias{finished}
\alias{status}
\alias{value}
\title{Status of an AsyncTask}
\usage{
status(...)
}
\arguments{
\item{...}{Arguments passed to the S3 method}
}
\value{
A character vector.
}
\description{
Status of an AsyncTask
}
\keyword{internal}

